Remove logs, move some staff to utils.
